// frontend/src/search-client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type FileCategory = 'image' | 'video' | 'audio' | 'document' | 'spreadsheet' | 'archive' | 'other';

export type HealthResponse = {
    status: string;
    index: string;
    engine: string;
};

export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type SearchResponse = {
    results?: Array<SearchResultItem>;
    next_cursor?: (string | null);
};

export type SearchResultItem = {
    id: string;
    name: string;
    folder_path: string;
    mime_type: string;
    category: FileCategory;
    size_bytes: number;
    created_at: string;
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
    input?: unknown;
    ctx?: {
        [key: string]: unknown;
    };
};

export type SearchFilesApiV1SearchFilesGetData = {
    category?: (FileCategory | null);
    cursor?: (string | null);
    folderPath: string;
    limit?: number;
    q?: (string | null);
};

export type SearchFilesApiV1SearchFilesGetResponse = (SearchResponse);

export type HealthCheckApiV1SearchHealthGetResponse = (HealthResponse);